Adding types file to database folder
